+++ lib/src/all_models.dart
@@ -1,5 +1,6 @@
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/serializer.dart';
+import 'package:built_value/standard_json_plugin.dart';
 
 // Import self in order to use in the metadata annotation.
 import 'all_models.dart';
@@ -22,6 +23,56 @@ export 'model/weapon.dart' show Weapon;
 
 part 'all_models.g.dart';
 
+class _CustomEnumJsonPlugin extends StandardJsonPlugin {
+  final Set<Type> _customEnumTypes;
+
+  _CustomEnumJsonPlugin(Iterable<Type> customEnumTypes)
+      : _customEnumTypes = customEnumTypes.toSet();
+
+  @override
+  Object beforeDeserialize(object, type) {
+    if (type.root == BuiltMap && type.parameters.isNotEmpty) {
+      if (_customEnumTypes.contains(type.parameters.first.root)) {
+        object = _addEnumEncoding(object as Map<String, Object>);
+      }
+    }
+    return super.beforeDeserialize(object, type);
+  }
+
+  Map<String, Object> _addEnumEncoding(Map<String, Object> json) {
+    return json.map((key, value) {
+      return MapEntry('"$key"', {
+        valueKey: value,
+        discriminator: '${value.runtimeType}',
+      });
+    });
+  }
+
+  @override
+  Object afterSerialize(object, type) {
+    var json = super.afterSerialize(object, type);
+    if (type.root == BuiltMap && type.parameters.isNotEmpty) {
+      if (_customEnumTypes.contains(type.parameters.first.root)) {
+        json = _removeEnumEncoding(json as Map<String, Object>);
+      }
+    }
+    return json;
+  }
+
+  Map<String, Object> _removeEnumEncoding(Map<String, Object> json) {
+    return json.map((key, value) {
+      var valOut = value;
+      if (key is String && key.startsWith('"')) {
+        key = key.substring(1, key.length - 1);
+      }
+      if (value is Map && value.containsKey(discriminator)) {
+        valOut = value[valueKey];
+      }
+      return MapEntry(key, valOut);
+    });
+  }
+}
+
 /// A [Serializers] for the models provided by this package.
 @SerializersFor([
   AttackDice,
@@ -39,6 +90,14 @@ part 'all_models.g.dart';
 ])
 final Serializers serializers = () {
   final builder = _$serializers.toBuilder()
+    ..addPlugin(_CustomEnumJsonPlugin(const [
+      AttackDice,
+      DefenseDice,
+      UnitKeyword,
+      UpgradeKeyword,
+      UpgradeSlot,
+      WeaponKeyword,
+    ]))
     ..add(Reference.serializer)
     ..add(UnitKeyword.serializer)
     ..add(UpgradeKeyword.serializer)